Change getComputerSelection to use switch case
This function previously used an else if statement to calculate the
selection for the computer where it randomized a number between 1 and 3
and then converted those numbers to variables based on the output. It
now still does this, but the conversion is made using a switch case to
make the code cleaner.